# web/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AuditEventView" : {
        "properties" : {
          "action" : {
            "type" : "string"
          },
          "actorId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "occurredAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "requestId" : {
            "type" : [ "string", "null" ]
          },
          "subject" : {
            "type" : "string"
          }
        },
        "required" : [ "action", "actorId", "id", "occurredAt", "requestId", "subject" ],
        "type" : "object"
      },
...
      "PageAuditEventView" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/AuditEventView"
            },
            "type" : "array"
          },
          "nextCursor" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "items", "nextCursor" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "sessionCookie" : {
        "description" : "Server-side session cookie. State-changing requests also need the X-XSRF-TOKEN header; call GET /api/auth/csrf first.",
        "in" : "cookie",
        "name" : "CERTFORGE_SESSION",
        "type" : "apiKey"
      }
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/admin/audit" : {
      "get" : {
        "operationId" : "events",
        "parameters" : [ {
          "in" : "query",
          "name" : "subject",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "actorId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "action",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "cursor",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageAuditEventView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "audit-controller" ]
      }
    },
...
  },
  "security" : [ {
    "sessionCookie" : [ ]
  } ]
}